requirements.txt
pytest
pytest-bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest